# crates/puddle-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "puddle",
    "description": "puddle's local API: pending requests, rules, audit, settings, consents and the event stream. Listens on 127.0.0.1 only. Every request needs `Authorization: Bearer <token>` (from the connection file), a `Host` of `127.0.0.1:<port>` or `localhost:<port>`, and no foreign `Origin`.",
    "contact": {
      "name": "Tijs van Kampen",
      "email": "puddle@tijsvankampen.be"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/rule-sets": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Every rule set (built-in first, then yours) and the System managed hosts with their reasons.",
        "operationId": "list_rule_sets",
        "responses": {
          "200": {
            "description": "the rule sets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetList"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Makes a rule set of your own, empty and on everywhere. Add entries with `POST /api/rules`\n(scope `set`) or by approving a request into it.",
        "operationId": "create_rule_set",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRuleSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "the new set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetView"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "name empty, too long or taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/rule-sets/{id}": {
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Renames a rule set of yours or changes its description.",
        "operationId": "update_rule_set",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "`user:<id>`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "the set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetView"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "a built-in set, or name empty, too long or taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Deletes a rule set of yours with its entries and switches.",
        "operationId": "delete_rule_set",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "`user:<id>`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the deleted set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetView"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "a built-in set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/rule-sets/{id}/switch": {
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Switches a set on or off for every workspace or for one (`enabled: null` removes that switch),\nand closes the open requests the set now decides.",
        "operationId": "switch_rule_set",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "`builtin:<slug>` or `user:<id>`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "the set and the requests it closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetSwitched"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "System managed has no switch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Every rule, including expired ones the sweeper hasn't removed yet.",
        "operationId": "list_rules",
        "responses": {
          "200": {
            "description": "the rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleList"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Creates a rule directly. Open requests it now decides are closed the same way.",
        "operationId": "create_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "the new rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "invalid pattern or expiry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, lower case, no trailing period. Never contains a token."
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "What went wrong, as a stable code a client can switch on.",
        "enum": [
          "misdirected_host",
          "forbidden_origin",
          "unauthorized",
          "bad_request",
          "unsupported_media_type",
          "payload_too_large",
          "invalid",
          "not_found",
          "conflict",
          "newer_settings",
          "unavailable",
          "internal"
        ]
      },
...
      "NewRuleRequest": {
        "type": "object",
        "description": "A rule to create directly (not from a pending request).",
        "required": [
          "scope",
          "pattern",
          "effect"
        ],
        "properties": {
          "effect": {
            "$ref": "#/components/schemas/Effect",
            "description": "Allow or deny."
          },
          "expires_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Epoch ms from which the rule no longer matches; left out or `null` is permanent.",
            "minimum": 0
          },
          "pattern": {
            "type": "string",
            "description": "`example.com` (exact), or `.example.com` / `*.example.com` (suffix; a public suffix is\nrefused)."
          },
          "scope": {
            "$ref": "#/components/schemas/RuleScope",
            "description": "Global or one workspace."
          }
        },
        "additionalProperties": false
      },
      "NewRuleSetRequest": {
        "type": "object",
        "description": "A rule set to make.",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "At most 500 characters."
          },
          "name": {
            "type": "string",
            "description": "1 to 64 characters, not used by another set."
          }
        },
        "additionalProperties": false
      },
...
      "Rule": {
        "type": "object",
        "description": "A rule.",
        "required": [
          "id",
          "scope",
          "pattern",
          "pattern_kind",
          "effect",
          "expires_at",
          "created_at",
          "created_by",
          "source_pending_id"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch ms.",
            "minimum": 0
          },
          "created_by": {
            "$ref": "#/components/schemas/Actor",
            "description": "Who created it."
          },
          "effect": {
            "$ref": "#/components/schemas/Effect",
            "description": "Allow or deny."
          },
          "expires_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Epoch ms from which the rule no longer matches; `null` is permanent.",
            "minimum": 0
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Row id, never reused."
          },
          "pattern": {
            "type": "string",
            "description": "`example.com` (exact) or `.example.com` (suffix)."
          },
          "pattern_kind": {
            "$ref": "#/components/schemas/PatternKind",
            "description": "Exact or suffix."
          },
          "scope": {
            "$ref": "#/components/schemas/RuleScope",
            "description": "Global or one workspace."
          },
          "source_pending_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The pending request it was decided from."
          }
        }
      },
...
      "RuleList": {
        "type": "object",
        "description": "Every rule, expired ones the sweeper hasn't removed yet included.",
        "required": [
          "rules"
        ],
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "description": "The rules."
          }
        }
      },
...
      "RuleSetEntry": {
        "type": "object",
        "description": "One entry of a rule set.",
        "required": [
          "pattern",
          "pattern_kind",
          "effect",
          "note",
          "rule_id",
          "expires_at"
        ],
        "properties": {
          "effect": {
            "$ref": "#/components/schemas/Effect",
            "description": "Allow or deny (built-in sets only allow)."
          },
          "expires_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Epoch ms from which the entry no longer counts; `null` is permanent.",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "description": "What it is for (built-in sets); empty for your own entries."
          },
          "pattern": {
            "type": "string",
            "description": "`example.com`, or `.example.com` for every name under it (a built-in entry is written\n`*.example.com`)."
          },
          "pattern_kind": {
            "$ref": "#/components/schemas/PatternKind",
            "description": "Exact or suffix."
          },
          "rule_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The entry's rule, for a set you made (delete or change it through `/api/rules`)."
          }
        }
      },
      "RuleSetKind": {
        "type": "string",
        "description": "Who made a rule set.",
        "enum": [
          "built_in",
          "user"
        ]
      },
      "RuleSetList": {
        "type": "object",
        "description": "Every rule set, and the System managed hosts.",
        "required": [
          "sets",
          "system_managed"
        ],
        "properties": {
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleSetView"
            },
            "description": "Built-in sets first, then yours by name."
          },
          "system_managed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemManagedHost"
            },
            "description": "The hosts puddle allows because of your setup, each with its reason. Your own rules\ndecide first: a deny of yours blocks one."
          }
        }
      },
      "RuleSetOverride": {
        "type": "object",
        "description": "One workspace's own switch for a set.",
        "required": [
          "workspace",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "On or off there."
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceName",
            "description": "The workspace."
          }
        }
      },
      "RuleSetSwitchRequest": {
        "type": "object",
        "description": "Switch a set on or off, for every workspace or for one.",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "On or off; `null` removes the switch, so the next level decides (the global switch, then\nthe set's default)."
          },
          "workspace": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/WorkspaceName",
                "description": "The workspace whose own switch to set; left out or `null` for every workspace."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "RuleSetSwitched": {
        "type": "object",
        "description": "What a switch did.",
        "required": [
          "set",
          "closed"
        ],
        "properties": {
          "closed": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Open requests the set now decides, closed the same way."
          },
          "set": {
            "$ref": "#/components/schemas/RuleSetView",
            "description": "The set as it is now."
          }
        }
      },
      "RuleSetUpdateRequest": {
        "type": "object",
        "description": "A rule set's new name and description.",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "At most 500 characters."
          },
          "name": {
            "type": "string",
            "description": "1 to 64 characters, not used by another set."
          }
        },
        "additionalProperties": false
      },
      "RuleSetView": {
        "type": "object",
        "description": "A rule set: where it is on, and its entries.",
        "required": [
          "id",
          "kind",
          "name",
          "description",
          "default_on",
          "global",
          "overrides",
          "entries",
          "changed_at",
          "created_at"
        ],
        "properties": {
          "changed_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Built-in sets: epoch ms when a puddle update last changed the entries; `null` if never.",
            "minimum": 0
          },
          "created_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Your sets: epoch ms it was made.",
            "minimum": 0
          },
          "default_on": {
            "type": "boolean",
            "description": "Whether it is on where nobody switched it: built-in sets ship off, your sets start on."
          },
          "description": {
            "type": "string",
            "description": "What it is for."
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleSetEntry"
            },
            "description": "Its entries."
          },
          "global": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "The switch for every workspace; `null` follows `default_on`."
          },
          "id": {
            "type": "string",
            "description": "`builtin:<slug>` or `user:<id>`."
          },
          "kind": {
            "$ref": "#/components/schemas/RuleSetKind",
            "description": "Built-in or yours."
          },
          "name": {
            "type": "string",
            "description": "Name."
          },
          "overrides": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleSetOverride"
            },
            "description": "Workspaces with a switch of their own."
          }
        }
      },
...
      "SystemManagedHost": {
        "type": "object",
        "description": "One host puddle allows because of your setup, and why.",
        "required": [
          "pattern",
          "note",
          "reason",
          "reason_text",
          "workspace"
        ],
        "properties": {
          "note": {
            "type": "string",
            "description": "What the host is for."
          },
          "pattern": {
            "type": "string",
            "description": "`example.com`, or `*.example.com` for every name under it."
          },
          "reason": {
            "$ref": "#/components/schemas/SystemReason",
            "description": "Why it is allowed."
          },
          "reason_text": {
            "type": "string",
            "description": "The reason in words."
          },
          "workspace": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/WorkspaceName",
                "description": "The one workspace it is allowed for; `null` for every workspace."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}